+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - `make_upgrade_patch.sh` helper for backporting framework improvements
 - Improvements to licensed asset encryption
 - Bump to use engine 0.9.0
+- Bugfix with cloning level repo already initialized with restricted assets
 
 ## 0.8.0 - 4/24/25